update integration test file
