+++ test/e2e/indexer/events/certification-events.spec.ts
